+++ src/Plugin.php
@@ -24,7 +24,7 @@ class Plugin extends \craft\base\Plugin
     /**
      * @inheritdoc
      */
-    public $schemaVersion = '1.0.1';
+    public $schemaVersion = '1.1';
 
 
     // Public Methods