+++ src/migrations/m180929_165000_remove_storageclass_setting.php
@@ -2,10 +2,12 @@
 
 namespace craft\awss3\migrations;
 
+use Craft;
 use craft\awss3\Volume;
 use craft\db\Migration;
 use craft\db\Query;
 use craft\helpers\Json;
+use craft\services\Volumes;
 
 /**
  * m180929_165000_remove_storageclass_setting migration.
@@ -17,25 +19,22 @@ class m180929_165000_remove_storageclass_setting extends Migration
      */
     public function safeUp()
     {
-        $volumes = (new Query())
-            ->select([
-                'id',
-                'settings',
-            ])
-            ->where(['type' => Volume::class])
-            ->from(['{{%volumes}}'])
-            ->all();
+        $projectConfig = Craft::$app->getProjectConfig();
+        $projectConfig->muteEvents = true;
 
+        foreach ($projectConfig->get(Volumes::CONFIG_VOLUME_KEY) as $uid => &$volume) {
+            if ($volume['type'] == Volume::class && isset($volume['settings']) && is_array($volume['settings'])) {
+                unset($volume['settings']['storageClass']);
 
-        foreach ($volumes as $volume) {
-            $settings = Json::decode($volume['settings']);
+                $this->update('{{%volumes}}', [
+                    'settings' => Json::encode($settings),
+                ], ['uid' => $uid]);
 
-            if ($settings !== null) {
-                unset($settings['storageClass']);
-                $settings = Json::encode($settings);
-                $this->update('{{%volumes}}', ['settings' => $settings], ['id' => $volume['id']]);
+                $projectConfig->set(Volumes::CONFIG_VOLUME_KEY . '.' . $uid, $volume);
             }
         }
+
+        $projectConfig->muteEvents = false;
     }
 
     /**